refactor(ui): BasicParamsGroup、StrategySettingsTabs、PreSendDialog 拆分为独立文件
