+++ src/danmaku_sender/ui/sender_page.py
@@ -1,15 +1,15 @@
 import logging
 
 from PySide6.QtWidgets import (
-    QWidget, QVBoxLayout, QHBoxLayout, QFormLayout, QLineEdit,
-    QPushButton, QComboBox, QLabel, QGroupBox, QTextEdit, QCheckBox,
-    QProgressBar, QTabWidget, QSpinBox, QDoubleSpinBox, QFrame,
-    QFileDialog, QMessageBox, QDialog
+    QWidget, QVBoxLayout, QHBoxLayout, QPushButton, QLabel, QDialog,
+    QGroupBox, QTextEdit, QProgressBar, QFileDialog, QMessageBox
 )
 from PySide6.QtGui import QTextCursor, QDragEnterEvent, QDropEvent
 from PySide6.QtCore import Qt, QDateTime, Signal, Slot
 
-from .framework.binder import UIBinder
+from .sender_basic_params import BasicParamsGroup
+from .sender_strategy_tabs import StrategySettingsTabs
+from .sender_dialogs import PreSendDialog
 from .framework.style_loader import SvgIcon
 
 from danmaku_sender.controller.video_controller import VideoController
@@ -474,345 +474,3 @@ class SenderPage(QWidget):
             event.acceptProposedAction()
 
 
-class BasicParamsGroup(QGroupBox):
-    """基础参数区"""
-    def __init__(self, state: AppState, parent=None):
-        super().__init__("基础参数", parent)
-        self.state = state
-        self._create_ui()
-
-    def _create_ui(self):
-        layout = QFormLayout(self)
-
-        # BV号
-        bv_layout = QHBoxLayout()
-        self.bv_input = QLineEdit()
-        self.bv_input.setPlaceholderText("请输入视频BV号")
-        self.fetch_btn = QPushButton("获取分P")
-        bv_layout.addWidget(self.bv_input)
-        bv_layout.addWidget(self.fetch_btn)
-
-        # 分P选择
-        self.part_combo = QComboBox()
-        self.part_combo.setPlaceholderText("请选择分P")
-        self.part_combo.setEnabled(False)
-
-        # 弹幕文件选择
-        file_layout = QHBoxLayout()
-        self.file_input = QLineEdit()
-        self.file_input.setReadOnly(True)
-        self.file_input.setPlaceholderText("请选择弹幕 XML 文件")
-        self.file_btn = QPushButton("选择文件")
-        file_layout.addWidget(self.file_input)
-        file_layout.addWidget(self.file_btn)
-
-        # 断点续传
-        self.skip_sent_cb = QCheckBox("启用断点续传 (跳过已发送)")
-        self.skip_sent_cb.setToolTip(
-            "开启后，发送前会自动检查数据库。\n"
-            "如果发现完全一致的弹幕（内容、时间、样式）已发送过，则自动跳过。"
-        )
-
-        layout.addRow(QLabel("视频BV号:"), bv_layout)
-        layout.addRow(QLabel("分P选择:"), self.part_combo)
-        layout.addRow(QLabel("弹幕文件:"), file_layout)
-        layout.addRow(QLabel(""), self.skip_sent_cb)
-
-    def init_bindings(self):
-        """将 UI 控件与 AppState 进行双向绑定"""
-        UIBinder.bind(self.bv_input, self.state.video_state, "bvid", realtime=True)
-        UIBinder.bind(self.skip_sent_cb, self.state.sender_config, "skip_sent")
-
-        if self.state.video_state.selected_part_name:
-            self.part_combo.setPlaceholderText(self.state.video_state.selected_part_name)
-
-    def set_inputs_locked(self, locked: bool):
-        """供主控调用的防误触锁"""
-        enabled = not locked
-
-        self.fetch_btn.setEnabled(enabled)
-        self.file_btn.setEnabled(enabled)
-        self.part_combo.setEnabled(enabled)
-        self.bv_input.setReadOnly(locked)
-        self.skip_sent_cb.setEnabled(enabled)
-
-
-class StrategySettingsTabs(QTabWidget):
-    """策略设置区"""
-    def __init__(self, state: AppState, parent=None):
-        super().__init__(parent)
-        self.state = state
-        self._create_ui()
-
-    def _create_ui(self):
-        # Tab 1: 发送设置
-        delay_tab = QWidget()
-        delay_layout = QHBoxLayout(delay_tab)
-        delay_layout.setContentsMargins(10, 20, 10, 20)
-
-        delay_layout.addWidget(QLabel("随机间隔(秒):"))
-        self.min_delay = QDoubleSpinBox()
-        self.min_delay.setRange(0.1, 60.0)
-        self.min_delay.setValue(8.0)
-        self.min_delay.setSingleStep(0.5)
-        delay_layout.addWidget(self.min_delay)
-
-        delay_layout.addWidget(QLabel("-"))
-
-        self.max_delay = QDoubleSpinBox()
-        self.max_delay.setRange(0.1, 60.0)
-        self.max_delay.setValue(8.5)
-        self.max_delay.setSingleStep(0.5)
-        delay_layout.addWidget(self.max_delay)
-
-        # 分隔线
-        delay_layout.addSpacing(15)
-        v_line = QFrame()
-        v_line.setFrameShape(QFrame.Shape.VLine)
-        v_line.setFrameShadow(QFrame.Shadow.Sunken)
-        delay_layout.addWidget(v_line)
-        delay_layout.addSpacing(15)
-
-        # 爆发模式
-        delay_layout.addWidget(QLabel("爆发模式: 每"))
-
-        self.burst_size = QSpinBox()
-        self.burst_size.setRange(0, 100)
-        self.burst_size.setToolTip("0 或 1 表示关闭爆发模式")
-        delay_layout.addWidget(self.burst_size)
-
-        delay_layout.addWidget(QLabel("条，休息"))
-
-        self.burst_rest_min = QDoubleSpinBox()
-        self.burst_rest_min.setRange(0.0, 300.0)
-        self.burst_rest_min.setValue(10.0)
-        self.burst_rest_min.setFixedWidth(60)
-        delay_layout.addWidget(self.burst_rest_min)
-
-        delay_layout.addWidget(QLabel("-"))
-
-        self.burst_rest_max = QDoubleSpinBox()
-        self.burst_rest_max.setRange(0.0, 300.0)
-        self.burst_rest_max.setValue(20.0)
-        self.burst_rest_max.setFixedWidth(60)
-        delay_layout.addWidget(self.burst_rest_max)
-
-        delay_layout.addWidget(QLabel("秒"))
-
-        delay_layout.addStretch()
-        self.addTab(delay_tab, "发送延迟")
-
-        # Tab 2: 自动终止
-        stop_tab = QWidget()
-        stop_layout = QHBoxLayout(stop_tab)
-        stop_layout.setContentsMargins(10, 20, 10, 20)
-
-        # 数量限制
-        stop_layout.addWidget(QLabel("已发送 >="))
-        self.stop_count = QSpinBox()
-        self.stop_count.setRange(0, 99999)
-        stop_layout.addWidget(self.stop_count)
-        stop_layout.addWidget(QLabel("条"))
-
-        stop_layout.addSpacing(20)
-        v_line2 = QFrame()
-        v_line2.setFrameShape(QFrame.Shape.VLine)
-        v_line2.setFrameShadow(QFrame.Shadow.Sunken)
-        stop_layout.addWidget(v_line2)
-        stop_layout.addSpacing(20)
-
-        # 时间限制
-        stop_layout.addWidget(QLabel("已用时 >="))
-        self.stop_time = QSpinBox()
-        self.stop_time.setRange(0, 99999)
-        stop_layout.addWidget(self.stop_time)
-        stop_layout.addWidget(QLabel("分钟"))
-
-        stop_layout.addStretch()
-        stop_layout.addWidget(QLabel("(0为不限制)"))
-
-        self.addTab(stop_tab, "自动终止")
-
-    def init_bindings(self):
-        """将 UI 控件与 AppState 进行双向绑定"""
-        config = self.state.sender_config
-
-        # 发送延迟策略
-        UIBinder.bind(self.min_delay, config, "min_delay")
-        UIBinder.bind(self.max_delay, config, "max_delay")
-
-        # 爆发模式
-        UIBinder.bind(self.burst_size, config, "burst_size")
-        UIBinder.bind(self.burst_rest_min, config, "rest_min")
-        UIBinder.bind(self.burst_rest_max, config, "rest_max")
-
-        # 自动终止规则
-        UIBinder.bind(self.stop_count, config, "stop_after_count")
-        UIBinder.bind(self.stop_time, config, "stop_after_time")
-
-    def set_inputs_locked(self, locked: bool):
-        """供主控调用的防误触锁"""
-        enabled = not locked
-
-        self.min_delay.setEnabled(enabled)
-        self.max_delay.setEnabled(enabled)
-        self.burst_size.setEnabled(enabled)
-        self.burst_rest_min.setEnabled(enabled)
-        self.burst_rest_max.setEnabled(enabled)
-        self.stop_count.setEnabled(enabled)
-        self.stop_time.setEnabled(enabled)
-
-
-class PreSendDialog(QDialog):
-    """发送前确认对话框：集中展示目标视频、弹幕统计、账号信息和发送策略"""
-
-    def __init__(self, state: AppState, parent=None):
-        super().__init__(parent)
-        self.state = state
-        # 使用临时拷贝，取消时不会污染全局配置
-        self._config_copy = state.sender_config.model_copy()
-
-        self.setWindowTitle("发送确认")
-        self.setMinimumWidth(420)
-        self.setWindowFlags(
-            Qt.WindowType.Dialog
-            | Qt.WindowType.WindowTitleHint
-            | Qt.WindowType.WindowCloseButtonHint
-        )
-
-        layout = QVBoxLayout(self)
-        layout.setSpacing(12)
-
-        # 1. 目标视频
-        video_group = QGroupBox("目标视频")
-        video_form = QFormLayout(video_group)
-        vs = state.video_state
-        video_form.addRow("BV号:", QLabel(vs.bvid))
-        video_form.addRow("标题:", QLabel(vs.video_title))
-        if vs.selected_part_name:
-            video_form.addRow("分P:", QLabel(vs.selected_part_name))
-        layout.addWidget(video_group)
-
-        # 2. 弹幕概览
-        danmaku_group = QGroupBox("弹幕信息")
-        danmaku_layout = QHBoxLayout(danmaku_group)
-        total = len(vs.loaded_danmakus)
-        valid = sum(1 for d in vs.loaded_danmakus if d.is_valid is not False)
-        invalid = sum(1 for d in vs.loaded_danmakus if d.is_valid is False)
-        danmaku_layout.addWidget(QLabel(f"总数: {total} 条"))
-        danmaku_layout.addWidget(QLabel(f"有效: {valid} 条"))
-        if invalid > 0:
-            warn_label = QLabel(f"无效: {invalid} 条")
-            warn_label.setStyleSheet("color: #e74c3c;")
-            danmaku_layout.addWidget(warn_label)
-        danmaku_layout.addStretch()
-        layout.addWidget(danmaku_group)
-
-        # 3. 当前账号
-        account_group = QGroupBox("当前账号")
-        account_layout = QHBoxLayout(account_group)
-        account_name = "未知"
-        for acc in state.saved_accounts:
-            if acc.sessdata == state.sessdata:
-                account_name = acc.name or "未命名"
-                break
-        sessdata = state.sessdata or ""
-        if len(sessdata) > 8:
-            masked = sessdata[:4] + "****" + sessdata[-4:]
-        elif sessdata:
-            masked = "****"
-        else:
-            masked = "(未登录)"
-        account_layout.addWidget(QLabel(f"{account_name} (SESSDATA: {masked})"))
-        account_layout.addStretch()
-        layout.addWidget(account_group)
-
-        # 4. 发送策略（绑定到临时拷贝，取消时自动丢弃）
-        strategy_group = QGroupBox("发送策略")
-        strategy_form = QFormLayout(strategy_group)
-        config = self._config_copy
-
-        # 延时
-        delay_layout = QHBoxLayout()
-        self.min_delay = QDoubleSpinBox()
-        self.min_delay.setRange(0.1, 60.0)
-        self.min_delay.setSingleStep(0.5)
-        delay_layout.addWidget(self.min_delay)
-        delay_layout.addWidget(QLabel("~"))
-        self.max_delay = QDoubleSpinBox()
-        self.max_delay.setRange(0.1, 60.0)
-        self.max_delay.setSingleStep(0.5)
-        delay_layout.addWidget(self.max_delay)
-        delay_layout.addWidget(QLabel("秒"))
-        delay_layout.addStretch()
-        strategy_form.addRow("随机间隔:", delay_layout)
-
-        # 爆发
-        burst_layout = QHBoxLayout()
-        self.burst_size = QSpinBox()
-        self.burst_size.setRange(0, 100)
-        burst_layout.addWidget(QLabel("每"))
-        burst_layout.addWidget(self.burst_size)
-        burst_layout.addWidget(QLabel("条，休息"))
-        self.burst_rest_min = QDoubleSpinBox()
-        self.burst_rest_min.setRange(0.0, 300.0)
-        self.burst_rest_min.setFixedWidth(60)
-        burst_layout.addWidget(self.burst_rest_min)
-        burst_layout.addWidget(QLabel("~"))
-        self.burst_rest_max = QDoubleSpinBox()
-        self.burst_rest_max.setRange(0.0, 300.0)
-        self.burst_rest_max.setFixedWidth(60)
-        burst_layout.addWidget(self.burst_rest_max)
-        burst_layout.addWidget(QLabel("秒"))
-        burst_layout.addStretch()
-        strategy_form.addRow("爆发模式:", burst_layout)
-
-        # 自动停止
-        stop_layout = QHBoxLayout()
-        self.stop_count = QSpinBox()
-        self.stop_count.setRange(0, 99999)
-        stop_layout.addWidget(QLabel("发满"))
-        stop_layout.addWidget(self.stop_count)
-        stop_layout.addWidget(QLabel("条"))
-        stop_layout.addSpacing(12)
-        self.stop_time = QSpinBox()
-        self.stop_time.setRange(0, 99999)
-        stop_layout.addWidget(QLabel("运行满"))
-        stop_layout.addWidget(self.stop_time)
-        stop_layout.addWidget(QLabel("分钟 (0=不限)"))
-        stop_layout.addStretch()
-        strategy_form.addRow("自动停止:", stop_layout)
-
-        # 断点续传
-        self.skip_sent_cb = QCheckBox("启用断点续传")
-        strategy_form.addRow("", self.skip_sent_cb)
-
-        layout.addWidget(strategy_group)
-
-        # 绑定到临时拷贝
-        UIBinder.bind(self.min_delay, config, "min_delay")
-        UIBinder.bind(self.max_delay, config, "max_delay")
-        UIBinder.bind(self.burst_size, config, "burst_size")
-        UIBinder.bind(self.burst_rest_min, config, "rest_min")
-        UIBinder.bind(self.burst_rest_max, config, "rest_max")
-        UIBinder.bind(self.stop_count, config, "stop_after_count")
-        UIBinder.bind(self.stop_time, config, "stop_after_time")
-        UIBinder.bind(self.skip_sent_cb, config, "skip_sent")
-
-        # 5. 按钮
-        btn_layout = QHBoxLayout()
-        btn_layout.addStretch()
-        cancel_btn = QPushButton("取消")
-        cancel_btn.clicked.connect(self.reject)
-        btn_layout.addWidget(cancel_btn)
-        confirm_btn = QPushButton("确认发送")
-        confirm_btn.setObjectName("primaryButton")
-        confirm_btn.clicked.connect(self._on_confirm)
-        btn_layout.addWidget(confirm_btn)
-        layout.addLayout(btn_layout)
-
-    def _on_confirm(self):
-        """确认时将临时拷贝写回全局配置"""
-        for field in self._config_copy.model_fields:
-            setattr(self.state.sender_config, field, getattr(self._config_copy, field))
-        self.accept()
